filecache: avoid caching entire file on random access
When reading large media files, software may often seek to
the end of the file for metadata.  This can be a large
performance penalty if fusedav must read the entire file
into temporary storage before it can read a few bytes
at the end.

Do not cache the entire file in this case, and do not mark
non-contiguous cached data as usable to other readers.