filecache: file_cache_sync_unlocked properly syncs entire file
When file_cache_sync_unlocked() calls load_up_to_unlocked(fi, -1)
the (l <= fi->present) check will incorrectly short-circuit
us out of loading the entire file.  So instead, we will will
use fi->length instead.

If we want to pass -1 to range.end for neon, we need to be
careful and not set range.end to any other negative value
because current versions of neon require (range.end == -1
(not: range.end < 0)) to indicate we want to read until
the end-of-file.