fix release() with multiple open() to the same path
dav_release() should not unlink a file from the cache until its
ref count drops to zero.  The ref count may be non-zero after
release because because dav_open() may be called multiple times
without a dav_release() in between, so relying on a boolean dead
flag is problematic.

Instead of relying on a boolean dead flag, rely on the ref count
of the file and only unlink it when the counter drops to zero.