filecache: avoid ABBA locking problem in unref
Since we have other code which locks files_mutex before
locking fi->mutex; we must preserve this ordering in
and not attempt to unlink wile fi->mutex is held.

Since file_cache_get will skip files on ref==0, it
is safe to leave the old file in the global file
cache with ref==0 after fi->mutex is unlocked.