setxattr: fix return values for errors
The return value for unsupported flags arguments must be
negative for FUSE.

Additionally, failure to parse an attribute name now returns
ENOTSUP instead of ENOATTR.  According to the setxattr(2)
manpage, ENOATTR is only for the XATTR_REPLACE flag.

Returning ENOTSUP for invalid attributes now allows GNU cp(1)
to successfully archive files with a "cp -a" invocation.